Added support vor fixed attachment with % and fixed issue with borders

+++ src/nodeparser.js
@@ -765,7 +765,7 @@ function isPositionedForStacking(container) {
     var position = container.css("position");
     var zIndex = (["absolute", "relative", "fixed"].indexOf(position) !== -1) ? container.css("zIndex") : "auto";
     // In chrome fixed has a default zIndex of 0 instead of auto
-    if(parseFloat(zIndex, 10) === 0) {
+    if(parseInt(zIndex, 10) === 0) {
         zIndex = "auto";
     }
     return zIndex !== "auto";
